fetch postCount as maxPosts and load into postCount; cleaned up createPost, updatePost, and deletePost
createPost and updatePost used to update PostListComponent immediately.
No longer. Also decided to simply start pagination on 0 not 1.
deletePost was changed to return the subscription and have the onDelete
in PostListComponent call getPosts() upon next. This is because unlike
createPost and updatePost, deletePost happens in the list of posts, not
on the post itself, and because returning the subscription allows using
the postService in PostList to call getPosts(), as opposed to making
some complicated way of getting posts in PostService directly.